Properly detect missing mandatory arguments to MI commands
Previously if an MI command had X mandatory and Y optional arguments you
could provide X or more optional arguments without providing any of the
mandatory arguments, and the argument validation code wouldn't complain.

For example this would pass argument validation even though the
mandatory `address` and `count` arguments are missing:

-data-read-memory-bytes --thread 1 --frame 0

Part of the problem was that an empty string was considered a valid
value for a mandatory argument, which didn't make much sense.